Rework the glance startup to handle db sync/version control
The glance server in essex must externally deal with the
version control and sync of the database.  This was done
in common, but we not modifying the database because the
database parms aren't available until after the api and
registry are configured.

Also add a sleep (hate this) in the glance setup function
to wait for the registry and api to start up.  Running
glance index isn't sufficient to ensure "up".