Fix initial call to glance-manage
Right now, on installation of glance, there will be a glance-manage
crash (visible in the glance-registry log). This is because we run
glance-manage in the api recipe, before glance-registry.conf is created
while glance-manage read settings from the glance-registry.conf file.

As it makes no sense to call glance-manage twice, we can simply stop
calling it from the api recipe and call it from the registry recipe; we
also make the registry recipe run before the api one.