Start at least two Glance API workers, limit it to four in total
When only one Glance API worker is running, long running operations
like streaming an image can block the API when it uses a native
backend that is not greenlet'ified (like for example rados) for
several minutes, depending on the image size. Start at least two
API workers to increase chance that the API is still available.