Services listen on another port when HA is enabled
When the HA proxy runs in front of the glance services (API and
registry), make sure that they bind to a different port.

In the config file templates of these services, the ports were set from
a config hash, so refactor this and pass the port as a template
variable from the recipes.

Also fix configs for scrubber and cache, as these use the registry
bind/port too.