Setup HA and VIP networks in crowbar framework
Check if HA should be enabled. Currently, that means that there are
clusters defined for a given role, i.e. that list of role elements
expands to a differing list of nodes.

Based on that, mark the role as using HA. Then set up virtual IP network
for each cluster and make sure the DNS is configured, as the setup
recipe in the glance cookbook needs to resolve virtual IPs in the cluster.