Fix glance-manage db sync happening twice on first chef run
glance-manage is looking at the config files of both glance-api and
glance-registry. When we run it, only glance-registry has been
configured and the connection string for database will be overridden by
the default one from glance-api, which points to sqlite.

This results in the migration happening on start of glance-api, which
breaks HA (since only one node can do the migration).

The fix is to run glance-manage with only the glance-registry config
file which we know is correct.

https://bugzilla.novell.com/show_bug.cgi?id=888584
(cherry picked from commit e1f0897a32ab4f6822d52f4f8a867122d14c7133)